+++ iac/azure/versions.tf
@@ -10,5 +10,9 @@ terraform {
       source  = "hashicorp/azuread"
       version = "~> 3.9"
     }
+    random = {
+      source  = "hashicorp/random"
+      version = "~> 3.9"
+    }
   }
 }